Merge 2017/12/03

+++ caffe2/operators/sparse_to_dense_mask_op.cc
@@ -1,3 +1,19 @@
+/**
+ * Copyright (c) 2016-present, Facebook, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include "caffe2/operators/sparse_to_dense_mask_op.h"
 
 namespace caffe2 {
@@ -14,7 +30,7 @@ OPERATOR_SCHEMA(SparseToDenseMask)
     .TensorInferenceFunction([](const OperatorDef& def,
                                 const vector<TensorShape>& in) {
       ArgumentHelper helper(def);
-      auto mask = helper.template GetRepeatedArgument<int>("mask");
+      auto mask = helper.template GetRepeatedArgument<int64_t>("mask");
       bool return_presence_mask = helper.template GetSingleArgument<bool>(
           "return_presence_mask", false);
       vector<TensorShape> out(1);
